requirements.txt
fastapi==0.70.0
google-cloud-pubsub
gunicorn==19.7.1
h11==0.12.0
h5py==2.10.0
httptools==0.2.0
Jinja2==2.10
jmespath==0.10.0
jsonpath-rw==1.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
libcst==0.3.21
livereload==2.6.1
lomond==0.3.3
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.3.4
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.1.0
pandas==1.0.5
Pillow==6.2.1
ply==3.11
prometheus-client==0.7.1
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil
python-dotenv==0.19.2
python-multipart==0.0.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.11.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
uvicorn==0.15.0
